change sa_flags in sigaction to int to match spec

See merge request redox-os/relibc!1016

+++ src/platform/linux/signal.rs
@@ -73,7 +73,7 @@ impl PalSignal for Sys {
         }
         let act = act.map(|act| {
             let mut act_clone = act.clone();
-            act_clone.sa_flags |= SA_RESTORER as c_ulong;
+            act_clone.sa_flags |= SA_RESTORER as c_int;
             act_clone.sa_restorer = Some(__restore_rt);
             act_clone
         });
